spring+redis+@cacheable

+++ src/main/java/ppf/dao/userMapper.java
@@ -17,7 +17,7 @@ public interface userMapper {
 
     int updateByPrimaryKey(User record);
 
-    User selectUser(long userId);
+    User selectUser(Integer userId);
 
     List<User>  selectAll();
